+++ next-docs/public/examples/avatarTW/Default.tsx
@@ -1,0 +1,5 @@
+import { Avatar } from '@heathmont/moon-core-tw';
+
+const Example = () => <Avatar />;
+
+export default Example;