+++ next-docs/public/examples/buttonTW/Default.tsx
@@ -1,0 +1,5 @@
+import { Button } from '@heathmont/moon-core-tw';
+
+const Example = () => <Button>Default</Button>;
+
+export default Example;