+++ next-docs/public/examples/checkboxTW/Default.tsx
@@ -1,10 +1,5 @@
 import { Checkbox } from '@heathmont/moon-core-tw';
 
-const Example = () => (
-  <Checkbox
-    label="I agree to receive bonus & marketing emails."
-    id="checkbox1"
-  />
-);
+const Example = () => <Checkbox />;
 
 export default Example;