+++ next-docs/public/examples/loaderTW/__tests__/index.test.tsx
@@ -1,0 +1,31 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import React from 'react';
+import { moonDesignLight, ThemeProvider } from '@heathmont/moon-themes';
+import renderer from 'react-test-renderer';
+import Default from '../Default';
+import Sizes from '../Sizes';
+import Colors from '../Colors';
+
+const renderWithTheme = (component: JSX.Element) => (
+  <ThemeProvider theme={moonDesignLight}>{component}</ThemeProvider>
+);
+
+describe('Loader', () => {
+  it('renders default', () => {
+    const testRenderer = renderer.create(renderWithTheme(<Default />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders sizes', () => {
+    const testRenderer = renderer.create(renderWithTheme(<Sizes />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders colors', () => {
+    const testRenderer = renderer.create(renderWithTheme(<Colors />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+});