+++ workspaces/themes/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @heathmont/moon-themes-tw
 
+## 9.1.7
+
+### Patch Changes
+
+- Fix build
+
+## 9.1.6
+
+### Patch Changes
+
+- Text input & auth code fixes
+
 ## 9.1.5
 
 ### Patch Changes